fix(admin): user deletion failed with 500 on notification cleanup
NotificationJpaRepository.deleteByRecipientId 声明为 long 返回类型，而
Spring Data 的 @Modifying JPQL 删除只允许 void/int，首次真实调用时抛
IllegalArgumentException 使删除用户接口返回 500。纯 Mockito 单测 mock 掉了
仓储无法暴露该签名问题。

- 返回类型改为 int（domain 接口与 JPA 实现同步）
- 新增 @DataJpaTest 切片测试用真实 Spring Data 执行链路防回归
- 已端到端验证：创建→删除普通用户成功，删除 docker-admin 被正确拒绝（403）

+++ server/skillhub-infra/src/main/java/com/iflytek/skillhub/infra/jpa/NotificationJpaRepository.java
@@ -46,5 +46,5 @@ public interface NotificationJpaRepository extends JpaRepository<Notification, L
     @Modifying
     @Transactional
     @Query("DELETE FROM Notification n WHERE n.recipientId = :recipientId")
-    long deleteByRecipientId(String recipientId);
+    int deleteByRecipientId(String recipientId);
 }
